Core: Update to 10.2.0
Port From (https://github.com/TrinityCore/TrinityCore/commit/1eea1d11b14d2ae065ab0f1888fba3b1ef03874f)

+++ Source/Game/DataStorage/Structs/A_Records.cs
@@ -146,7 +146,7 @@ namespace Game.DataStorage
         public Vector3 Pos;
         public uint Id;
         public ushort ContinentID;
-        public sbyte PhaseUseFlags;
+        public int PhaseUseFlags;
         public ushort PhaseID;
         public ushort PhaseGroupID;
         public float Radius;
